+++ crates/outlayer/sdk/src/host/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(target_family = "wasm")]
+pub mod sys;
