+++ crates/outlayer/sdk/src/host/sys/mod.rs
@@ -1,0 +1,6 @@
+pub struct SysHost;
+
+#[cfg(feature = "ed25519")]
+mod ed25519;
+#[cfg(feature = "secp256k1")]
+mod secp256k1;
