+++ crates/outlayer/sdk/src/lib.rs
@@ -1,12 +1,7 @@
-// #[cfg(feature = "ed25519")]
-// pub mod ed25519;
-// // #[cfg(feature = "secp256k1")]
-// pub mod secp256k1;
+mod host;
 
-// #[cfg(target_family = "wasm")]
-pub mod sys;
+#[cfg(target_family = "wasm")]
+pub type Host = host::sys::SysHost;
 
-// #[cfg(target_family = "wasm")]
-pub type Host = sys::SysHost;
-// #[cfg(not(target_family = "wasm"))]
-// pub type Host = defuse_outlayer_host::DefaultHost;
+#[cfg(not(target_family = "wasm"))]
+pub type Host = defuse_outlayer_host::DefaultHost;
