+++ crates/outlayer/sys/src/ed25519.rs
@@ -1,0 +1,6 @@
+wit_bindgen::generate!({
+    path: "../wit/crypto.wit",
+    world: "ed25519-world",
+});
+
+pub use outlayer::host::ed25519;
